lib: Pass full path for RIB entries
The read functions for the RIB will now receive the full path, instead
of only the entry name. For IPCPs, we organized the RIB in an

/<ipcp>/<component>/entries

structure with a directory per component, so we don't need the full
path at this point. For process flow information, it's a lot more
convenient to organize it the following way

/<pid>/<fd>/stat

We can then register/unregister the flow descriptor when the frct
instance is created, and for getting the stats, we'd know the flow
descriptor from the fuse file path. If we would create a file per flow
instead of a directory per flow, something like

/<pid>/flows/<fd>

we'd need to do additional bookkeeping to list the contents of that
directory (we would need to track all flows with an active FRCT
instance), that fuse knows because it tracks the directories.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>