lib: Don't use pointer to set FRCT flags
The fccntl call FRCTSFLAGS was using a pointer to a flags so set
flags, which should just be a regular uint16_t.

For instance, the FRCTLINGER flags can now be turned off using

fccntl(fd, FRCTSFLAGS, FRCTFRESCNTL | FRCTFRTX)

leaving only resource control (flow control, FRCTFRESCNTL) and
retransmission enabled. Note that retransmission (FRCTFRTX) can't be
enabled or disabled on a live flow, it will be set on flow allocation.

Updates the man page for fccntl to add these FRCT options.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>