lib: Don't log prefix in syslog
Logging the prefix in the system logs has a lot of duplication, as the
syslog already includes the name of the daemon. Maybe we should
deprecate logging to stdout and focus on the syslog, revise things a
bit to print internal component names if they are defined. For now, I
think it's less annoying to read the syslog without the prefixes.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>