ipcpd: Refactor main functions
The structure of main functions of the IPCPs was a bit strange with a
ipcp_shutdown() call that combined waiting for a terminating signal
with stopping the internal threads. This is now revised into a
symmetrical design of

ipcp_start(), which now includes the create response towards the IRMd.

ipcp_sigwait(), which waits for a shutdown signal

ipcp_stop() that then stops the internal threads.

Now the main() functions of the IPCPs will make sense without checking
what that ipcp_shutdown() functions actually does.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>