icpdp: Fix layer information for ipcpd-udp
The refactors removed the need to set the hash algorithm for the
ipcpd-udp and the ipcpd-broadcast. However, the algorithm was not set
at bootstrap, so the ipcpd-udp was trying to use an SHA3-256 instead
of an MD5, causing flow allocation over the UDP to fail. The
ipcpd-broadcast used the default, so there was no problem.

Fixed by setting the correct algorithm for these ipcpds at bootstrap.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>