include: Add option to prefix logs with an ID
This adds the log_xxx_id() macros to print a logline prefixed with a
64-bit identifier in hex format, for instance:

  [2a92d78c2f52b76a] Hello ouroboros!

The ID is assumed to be 8 bytes in big endian format (it uses the
HASH_FMT functions and hashes are stored in big endian byte arrays).

The implementation uses the compiler-specific '##' operator to allow
empty variadics. E.g.

	func(arg , ## __VA_ARGS__);

will eat the comma if __VA_ARGS__ is empty and thus allow

     	func(arg);

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>