ipcpd: Simplify including common enrollment code
The common enrollment code between the unicast and broadcast IPCP
included the same C source from a source file in the unicast and
broadcast directories. Now it's handled by defining COMMON_SOURCES for
the unicast and broadcast IPCP.

For now, only the enrollment component is common. Two things are
needed for the connection manager:

       * Routing to be fixed to use a supporting broadcast layer. Then
         management flows can be removed from the unicast IPCP (as
         they are in fact data transfer flows of the broadcast layer
         that supports the routing dissemination traffic).

       * DHT to enroll as part of IPCP enrollment. Now it enrolls when
         a data transfer flow is established to be more akin to how
         Kademlia nodes join a Kademlia DHT, but this should be
         revised to adhere more to O7s enrollment concepts.

After that, the connection manager code will also be completely shared
between the unicast and broadcast layer and the connmgr will also be
common code.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>