irmd: Fix parsing multiple args in configfile
New method retains the original string in parsing the args string into
an argv dynamic array.  Previous method (`strtok`) didn't work,
because it is a destructive function, changing the supplied string. We
however needed to apply it twice to the same string.  It is still done
twice in a loop, to make sure argc is exact. Other methods, like
counting the amount of spaces to determine argc, would be incorrect
for his particular way of tokenizing if arguments are separated by
e.g. two spaces.

Also fixes a wrong pointer dereference, which did go unnoticed before
due to the previous error.

Signed-off-by: Thijs Paelman <thijs@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>