irmd: Fix parsing empty argument in configfile
Previously, args=[""] would crash the IRMd when it tried to
auto-instantiate the program.
Now, specifying an empty string as argument to a program WILL
automatically start that program if there is an incoming flow,
but without any arguments, as intended.

Remark: The semantics might be a bit less clear than
e.g. args=["--listen"], but they are a natural extension.

Signed-off-by: Thijs Paelman <thijs@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>