irmd: Error when config not found
Error instead of warn when the --config param is set, but the file is
not found. This prevents the IRMd from being in a 'half state', where it
is started, but the expected config is not applied.
One simply has to remove the --config option OR supply a correct
filepath.

Also fix compile option for the help message.

Signed-off-by: Thijs Paelman <thijs@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>