build: Fix build and tests on appveyor
The configuration parameter was not correctly passed, which caused the
build system to build with debug symbols (configuration does not match
"Release" and not set CONFIG_OUROBOROS_DEBUG, causing the compiler to
include the debugging code. Fixed appveyor command. The build will now
fail on an unknown CMAKE_BUILD_TYPE.

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>