build: Fix tests with CMake >= 3.29
Removing the testdriver source by the driver name doesn't work anymore
in CMake 3.29 because of the following (breaking) change:

Changed in version 3.29: The test driver source is listed by absolute
path in the build tree. Previously it was listed only as <driverName>.

https://cmake.org/cmake/help/latest/command/create_test_sourcelist.html

When using CMake 3.29 or above, Ouroboros will use the list POP_FRONT
function (introduced in CMake 3.15) to get rid of it.

https://cmake.org/cmake/help/latest/command/list.html#pop-front

Signed-off-by: Dimitri Staessens <dimitri@ouroboros.rocks>
Signed-off-by: Sander Vrijders <sander@ouroboros.rocks>